Hack timezone support in
Explicitly set the Pacific timezone on the various dates and then
hack the calendar display to show the right day of the week.

This hack has the side effect of starting the week on Monday for
those of us that live on the left side of the International Date
Line. A nice bonus :)